app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
alembic==1.8.1
attrs==22.1.0
azure-ai-formrecognizer==3.2.0b2
azure-common==1.1.28
azure-core==1.22.1
backoff==1.11.1
beautifulsoup4==4.11.1
blis==0.7.8
bs4==0.0.1
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cymem==2.0.6
databricks-cli==0.17.0
dill==0.3.5.1
docker==5.0.3
docopt==0.6.2
elastic-apm==6.10.1
elasticsearch==7.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
farm-haystack==1.6.0
filelock==3.7.1
Flask==2.2.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.12.0
inflect==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.0
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
mlflow==1.27.0
mmh3==3.0.0
monotonic==1.6
more-itertools==8.13.0
msrest==0.6.21
murmurhash==1.0.7
networkx==2.8.5
nltk==3.7
num2words==0.5.10
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathy==0.6.2
Pillow==9.2.0
posthog==1.4.9
preshed==3.0.6
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
protobuf==4.21.4
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
PyYAML==6.0
quantulum3==0.7.10
querystring-parser==1.2.4
rapidfuzz==2.4.2
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
SQLAlchemy==1.4.39
sqlparse==0.4.2
srsly==2.4.4
tabulate==0.8.10
thinc==8.1.0
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.12.0+cpu
torchaudio==0.12.0+cpu
torchvision==0.13.0+cpu
tqdm==4.64.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
websocket-client==1.3.3
Werkzeug==2.2.1
zipp==3.8.1
